Move week changer into its own component.
