+++ src/GoatSlips/Client/src/components/WeekChanger.module.scss
@@ -1,0 +1,19 @@
+@import '../variables';
+
+.weekChanger {
+    display: flex;
+    justify-content: space-between;
+    padding: 0.2em;
+    align-items: center;
+}
+
+.weekChangeInput {
+    margin-left: 0.2em !important;
+    margin-right: 0.2em !important;
+}
+
+.squareIconButton {
+    width: $headerHeight;
+    height: $headerHeight;
+    border-radius: 4px !important;
+}
