+++ src/GoatSlips/Client/src/helpers/dateHelpers.ts
@@ -1,0 +1,6 @@
+export const getSundayDateForDate = (date: Date) => {
+    const currentDay = date.getDay();
+    const calculatedSundayDate = new Date(date);
+    calculatedSundayDate.setDate(calculatedSundayDate.getDate() - currentDay);
+    return calculatedSundayDate;
+};
